[webgpu] Simplify shader key of binary op and unary op
