+++ tfjs-backend-webgpu/src/kernels/Abs.ts
@@ -18,10 +18,10 @@
 import {Abs, KernelConfig} from '@tensorflow/tfjs-core';
 import {unaryKernelFunc} from '../kernel_utils/kernel_funcs_utils';
 import {simpleAbsImplCPU} from '../kernel_utils/shared';
-import {ABS} from './unary_op_webgpu';
+import {UnaryOpType} from './unary_op_util';
 
 export const abs =
-    unaryKernelFunc({opSnippet: ABS, cpuKernelImpl: simpleAbsImplCPU});
+    unaryKernelFunc({opType: UnaryOpType.ABS, cpuKernelImpl: simpleAbsImplCPU});
 
 export const absConfig: KernelConfig = {
   kernelName: Abs,
