+++ tfjs-backend-webgpu/src/kernels/Equal.ts
@@ -20,7 +20,7 @@ import {Equal, KernelConfig} from '@tensorflow/tfjs-core';
 import {binaryKernelFunc} from '../kernel_utils/kernel_funcs_utils';
 import {equalImplCPU as cpuEqual} from '../kernel_utils/shared';
 
-import {BinaryOpType} from './binary_ops';
+import {BinaryOpType} from './binary_op_util';
 
 export const equal = binaryKernelFunc(
     {opSnippet: BinaryOpType.EQUAL, dtype: 'bool', cpuKernelImpl: cpuEqual});
