+++ tfjs-backend-webgpu/src/kernels/Exp.ts
@@ -18,9 +18,10 @@
 import {Exp, KernelConfig} from '@tensorflow/tfjs-core';
 import {unaryKernelFunc} from '../kernel_utils/kernel_funcs_utils';
 import {expImplCPU} from '../kernel_utils/shared';
-import {EXP} from './unary_op_webgpu';
+import {UnaryOpType} from './unary_op_util';
 
-export const exp = unaryKernelFunc({opSnippet: EXP, cpuKernelImpl: expImplCPU});
+export const exp =
+    unaryKernelFunc({opType: UnaryOpType.EXP, cpuKernelImpl: expImplCPU});
 
 export const expConfig: KernelConfig = {
   kernelName: Exp,
