+++ tfjs-backend-webgpu/src/kernels/FloorDiv.ts
@@ -17,7 +17,7 @@
 
 import {FloorDiv, KernelConfig} from '@tensorflow/tfjs-core';
 import {binaryKernelFunc} from '../kernel_utils/kernel_funcs_utils';
-import {BinaryOpType} from './binary_ops';
+import {BinaryOpType} from './binary_op_util';
 
 export const floorDiv =
     binaryKernelFunc({opSnippet: BinaryOpType.INT_DIV, dtype: 'int32'});
