+++ tfjs-backend-webgpu/src/kernels/GreaterEqual.ts
@@ -18,7 +18,7 @@
 import {GreaterEqual, KernelConfig} from '@tensorflow/tfjs-core';
 import {binaryKernelFunc} from '../kernel_utils/kernel_funcs_utils';
 import {greaterEqualImplCPU as cpuGreaterEqual} from '../kernel_utils/shared';
-import {BinaryOpType} from './binary_ops';
+import {BinaryOpType} from './binary_op_util';
 
 export const greaterEqual = binaryKernelFunc({
   opSnippet: BinaryOpType.GREATER_EQUAL,
