+++ tfjs-backend-webgpu/src/kernels/Less.ts
@@ -18,7 +18,7 @@
 import {KernelConfig, Less} from '@tensorflow/tfjs-core';
 import {binaryKernelFunc} from '../kernel_utils/kernel_funcs_utils';
 import {lessImplCPU as cpuLess} from '../kernel_utils/shared';
-import {BinaryOpType} from './binary_ops';
+import {BinaryOpType} from './binary_op_util';
 
 export const less = binaryKernelFunc(
     {opSnippet: BinaryOpType.LESS, dtype: 'bool', cpuKernelImpl: cpuLess});
