+++ tfjs-backend-webgpu/src/kernels/LessEqual.ts
@@ -18,7 +18,7 @@
 import {KernelConfig, LessEqual} from '@tensorflow/tfjs-core';
 import {binaryKernelFunc} from '../kernel_utils/kernel_funcs_utils';
 import {lessEqualImplCPU as cpuLessEqual} from '../kernel_utils/shared';
-import {BinaryOpType} from './binary_ops';
+import {BinaryOpType} from './binary_op_util';
 
 export const lessEqual = binaryKernelFunc({
   opSnippet: BinaryOpType.LESS_EQUAL,
