+++ tfjs-backend-webgpu/src/kernels/LogicalAnd.ts
@@ -19,7 +19,7 @@ import {KernelConfig, LogicalAnd} from '@tensorflow/tfjs-core';
 
 import {binaryKernelFunc} from '../kernel_utils/kernel_funcs_utils';
 
-import {BinaryOpType} from './binary_ops';
+import {BinaryOpType} from './binary_op_util';
 
 export const logicalAnd = binaryKernelFunc({
   opSnippet: BinaryOpType.LOGICAL_AND,
