+++ tfjs-backend-webgpu/src/kernels/Minimum.ts
@@ -20,7 +20,7 @@ import {KernelConfig, Minimum} from '@tensorflow/tfjs-core';
 import {binaryKernelFunc} from '../kernel_utils/kernel_funcs_utils';
 import {minimumImplCPU as cpuMinimum} from '../kernel_utils/shared';
 
-import {BinaryOpType} from './binary_ops';
+import {BinaryOpType} from './binary_op_util';
 
 export const minimum = binaryKernelFunc({
   opSnippet: BinaryOpType.MIN,
