+++ tfjs-backend-webgpu/src/kernels/Multiply.ts
@@ -18,7 +18,7 @@
 import {KernelConfig, Multiply} from '@tensorflow/tfjs-core';
 import {binaryKernelFunc} from '../kernel_utils/kernel_funcs_utils';
 import {multiplyImplCPU as cpuMultiply} from '../kernel_utils/shared';
-import {BinaryOpType} from './binary_ops';
+import {BinaryOpType} from './binary_op_util';
 
 export const multiplyKernelFunc = binaryKernelFunc({
   opSnippet: BinaryOpType.MUL,
