+++ tfjs-backend-webgpu/src/kernels/NotEqual.ts
@@ -20,7 +20,7 @@ import {KernelConfig, NotEqual} from '@tensorflow/tfjs-core';
 import {binaryKernelFunc} from '../kernel_utils/kernel_funcs_utils';
 import {notEqualImplCPU as cpuNotEqual} from '../kernel_utils/shared';
 
-import {BinaryOpType} from './binary_ops';
+import {BinaryOpType} from './binary_op_util';
 
 export const notEqual = binaryKernelFunc({
   opSnippet: BinaryOpType.NOT_EQUAL,
