+++ tfjs-backend-webgpu/src/kernels/Pow.ts
@@ -18,7 +18,7 @@
 import {KernelConfig, Pow} from '@tensorflow/tfjs-core';
 
 import {binaryKernelFunc} from '../kernel_utils/kernel_funcs_utils';
-import {BinaryOpType} from './binary_ops';
+import {BinaryOpType} from './binary_op_util';
 
 export const pow = binaryKernelFunc({
   opSnippet: BinaryOpType.POW,
