+++ tfjs-backend-webgpu/src/kernels/Relu6.ts
@@ -17,9 +17,9 @@
 
 import {KernelConfig, Relu6} from '@tensorflow/tfjs-core';
 import {unaryKernelFunc} from '../kernel_utils/kernel_funcs_utils';
-import {RELU6} from './unary_op_webgpu';
+import {UnaryOpType} from './unary_op_util';
 
-export const relu6 = unaryKernelFunc({opSnippet: RELU6});
+export const relu6 = unaryKernelFunc({opType: UnaryOpType.RELU6});
 
 export const relu6Config: KernelConfig = {
   kernelName: Relu6,
