+++ tfjs-backend-webgpu/src/kernels/Square.ts
@@ -17,7 +17,8 @@
 
 import {KernelConfig, Square, SquareInputs} from '@tensorflow/tfjs-core';
 import {WebGPUBackend} from '../backend_webgpu';
-import {SQUARE, UnaryOpProgram} from './unary_op_webgpu';
+import {UnaryOpProgram} from './unary_op_webgpu';
+import {UnaryOpType} from './unary_op_util';
 
 export const squareConfig: KernelConfig = {
   kernelName: Square,
@@ -25,7 +26,7 @@ export const squareConfig: KernelConfig = {
   kernelFunc: ({inputs, backend}) => {
     const {x} = inputs as SquareInputs;
     const webGPUBackend = backend as WebGPUBackend;
-    const program = new UnaryOpProgram(x.shape, SQUARE);
+    const program = new UnaryOpProgram(x.shape, UnaryOpType.SQUARE);
     return webGPUBackend.runWebGPUProgram(program, [x], x.dtype);
   }
 };
