+++ tfjs-backend-webgpu/src/kernels/activation_util.ts
@@ -17,25 +17,25 @@
 
 import {backend_util} from '@tensorflow/tfjs-core';
 
-import {BinaryOpType, getBinaryOpString} from './binary_ops';
-import * as unary_op from './unary_op_webgpu';
+import {BinaryOpType, getBinaryOpString} from '../kernels/binary_op_util';
+import {getUnaryOpString, UnaryOpType} from '../kernels/unary_op_util';
 
 export function mapActivationToShaderProgram(
     activation: backend_util.Activation, packed = false): string {
   if (activation === null) {
     return null;
   } else if (activation === 'linear') {
-    return unary_op.LINEAR;
+    return getUnaryOpString(UnaryOpType.LINEAR);
   } else if (activation === 'relu') {
-    return packed ? unary_op.RELU_VEC4 : unary_op.RELU;
+    return getUnaryOpString(UnaryOpType.RELU, packed);
   } else if (activation === 'elu') {
-    return packed ? unary_op.ELU_VEC4 : unary_op.ELU;
+    return getUnaryOpString(UnaryOpType.ELU, packed);
   } else if (activation === 'relu6') {
-    return unary_op.RELU6;
+    return getUnaryOpString(UnaryOpType.RELU6);
   } else if (activation === 'prelu') {
     return getBinaryOpString(BinaryOpType.PRELU, packed);
   } else if (activation === 'sigmoid') {
-    return unary_op.SIGMOID;
+    return getUnaryOpString(UnaryOpType.SIGMOID);
   }
   throw new Error(`Activation ${
       activation} has not been implemented for the WebGPU backend.`);
